backend/requirements.txt
flask
flask-sqlalchemy
flask-smorest
flask-cors
marshmallow
python-dotenv
PyJWT
psycopg[binary]
